requirements.txt
asgiref==3.8.1
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
pywin32-ctypes==0.2.3
rapidfuzz==3.9.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.1
tomlkit==0.13.2
trove-classifiers==2024.9.12
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.5
